transformation/ml-model/requirements.txt
google-cloud-bigquery==3.25.0
python-dotenv==1.0.0
